subscribe once to global add field
