+++ contrib/occtax/frontend/app/occtax-form/counting/counting.component.html
@@ -1,3 +1,4 @@
+{{countingForm.value |json}}
 <div class="row">
   <div
     *ngIf="occtaxConfig.form_fields.obj_count"
